codathon-phase3/requirements.txt
FastAPI
llama_cpp_python
pypdf
sentence_transformers
chromadb
uvicorn